travis-ci: fix fluxorama docker tag
Problem: A shell variable expansion bug in the travis config
results in failure to build and push a version of fluxorama
docker image with each merge to the main Flux branch.

Simplify the shell script in .travis.yml that builds and pushes
the fluxorama image, an in the process fix the shell expansion
bug.