flux-job: Add totalview_jobid support
Problem: Debuggers such as TotalView and STAT require
a symbol named totalview_jobid in the process address
space of the starter process to fetch and to use this
target jobid for bulk tool daemon launching.

Add totalview_jobid of char* type and fill
it with the jobid of C-string when MPIR_being_debug
is set.