flux-job: add --debug to allow for debugger attach using PID
Problem: Debuggers behave much better when they attach to a
running job by attaching to a starter process.
Currently, this can be achieved by starting flux job attach
with the --debug-emulate switch to become the starter
process that debuggers interface,
but it is a hidden option and perhaps more importatnly
this also has a side effect:
sending an additional SIGCONT to the job
to faciliate our testing.

Add --debug flag as user visible option.

Make it enable MPIR in the same way as --debug-emulate
but does not send the additional SIGCONT to
the job.