shell: rc: error if no handlers provided in plugin.register
Problem: If plugin.register is used in the shell's initrc, and
handlers is accidentally defined as a table instead of an array,
a plugin with no handlers is silently registered with the shell.
This can be very difficult to diagnose, because no callbacks will
ever be called.

Assume it is an error if plugin.register() is called without
any handlers. Generate an error message with a hint that the handlers
argument should be an array of tables, not a table itself.