broker/state_machine: add JOIN state
Add a new JOIN state, in which brokers wait for their parent
to become ready before entering the next state.  In this initial
commit, JOIN simply waits for the parent to enter JOIN state.
In the future it will wait for the parent to complete rc1.

Add a streaming RPC that permits a broker's state transitions
to be monitored.  The RPC responses can cause parent-ready or
parent-fail events to be posted, and also provides a mechanism
for tracking parent exit from the BARRIER state, to be added
in a subsequent commit.