flux-job: initialize locale at program startup
Problem: flux-job does not initialize locale and thus always runs
with effective locale set to "C". This doesn't allow the program to
take advantage of the special unicode character in the F58 JOBID
prefix when running in a UTF-8 capable locale.

Call

 setlocale (LC_ALL, "")

at program startup to initialize locale from system configuration
and/or the user's environment.