flux-mini: add environment manipulation arguments
Problem: flux-mini always copies the entire current environment into
jobspec, and there is no way to filter, modify, or amend it.

Add a new function get_filtered_environment() to flux-mini.py which
accepts and array of environment modification "rules" to apply to the
current process environment in order to form the final environment
copied into the jobspec.

The rules are

 * If a rule begins with '-', then the rest of the rule is a pattern
   which removes matching environment variables. If the pattern
   starts with '/', it is a regex (optionally ending with '/'),
   otherwise the pattern is considered a shell glob(7).

    Example: "-*" filters all variables, creating an empty environment.

 * If a rule begins with '^' then the rest of the rule is a filename
   from which to read more rules, one per line. The ~ character is
   expanded in filename to be users directory.

    Example: "^~/envfile" reads from file $HOME/envfile.

 * If a rule is of the form `VAR=VAL`, then the variable VAR is set
   to VAL. Before being set, however, variables in VAL are expanded
   via string.Template().substitute, so any $VARIABLE within VAL will
   be expanded using first the generated environment, falling back
   to os.environ.

     Example: "PATH=/bin", "PATH=$PATH:/bin"

 * Otherwise, the rule is considered a pattern from which to match
   variables from os.environ if they do not alread exist in the
   generated environment. E.g. "PATH" will export PATH from the
   current environment (if it has not already been set in the
   generated environment), and "OMP*" would copy all environment
   variables that start with "OMP" and are not already set in the
   generated environment.

     Example: "PATH", "FLUX_*_PATH"

Rules are specified on the command line with a new `--env` option.
For convenience, `--env-remove=PATTERN` and `--env-file=FILE` are also
provided, which are equivalent to `--env=-PATTERN` and `--env=^FILE`,
respectively. All rules are applied in the order in which they are
specified.

The first implied rule is always `--env=*`, i.e. the generated
environment always starts as a copy of the current environment.

Since an env file is just a set of rules, it can be used in place of
multiple command line arguments, and can even load other env files:

-FLUX_URI
-LS_COLORS
-OMPI_*
PATH=$PATH:/some/other/path
^~/.flux/some-other-env-file