shell: always initialize jobspec->options to empty object
Problem: If no shell options are set in jobspec, then jobspec->options
is NULL instead of an empty object. This requires special handling in
the flux_shell_setopt() functions, so it would be better if an empty
json object was used instead.

If jobspec->options is NULL after parsing the jobspec object, initialize
to an empty object.

Fixes #3158