shell: gpubind: prevent jobs from using GPUs unless allocated
Problem: Jobs that are executed on nodes with GPUs, but have not
allocated those GPUs can still run on the GPUs since CUDA_VISIBLE_DEVICES
is not set when the number of allocated GPUs is 0.

Set a default value for CUDA_VISIBLE_DEVICES=-1, which should prevent
any processes in the job from seeing available GPUs on a node. This
behavior is disabled if the gpu-affinity=off shell option is specified.

Fixes #3154