shell: allow default initrc path to be set via broker attr
Problem: There is no way to override the default shell initrc path
for an instance, on per-job via the initrc shell jobspec option.

Instead of an environment variable to set a default shell initrc
path, allow the initrc path to be set by a broker attribute. The
default path may then be overridden on a per-instance basis on the
commandline or via rc1 script, and is therefore tied to the instance
instead of being set and propagated in the environment.