job-info: drop fatal jobspec parse error to a warning
Problem: Non-V1 jobspec causes a fatal error in job-info when it is
calculating the number of tasks. This prevents the job from reaching
Fluxion, even if Fluxion would be capable of scheduling it and the
job-exec would be capable of running it based on the emitted R-lite.

Solution: Drop the hard error down to a warning and set slot count to
-1, which results in a task count of -1.