shell: implement recursive parser for non-V1 jobspec
Problem: the job-shell fatally errors out when parsing non-V1
jobspec. In particular, jobspec that is very similar to V1 jobspec but
contains extra resources above the `core` level causes issues.

Solution: implement a more complex recursive parser for jobspec within
the job-shell.  This parser requires that the jobspec resource ordering
is the same as the ordering specified in V1, but it allows additional
resources before and in between the V1 resources (i.e., node, slot, and
core).  In shorthand, it requires that the jobspec follows the form
`...->[node]->...->slot->...->core`.  Where `node` is optional, and
`...` represents any non-V1 resources. Note that this more complex
parser is only used as a fallback when the V1-specific hardcoded parsing
fails.