doc: flux-shell(1): expand flux_plugin_init() section
Problem: The flux-shell describes the use of flux_plugin_init(3) for
dynamic plugins as a spot to only register callbacks for other shell
topic strings. However, this "init" callback can be used for more
than just that.

Document that flux_plugin_init() is not called for shell builtins,
and is therefore a spot where DSO shell plugins may influence the
operation of builtin plugins before "shell.init".