job-shell: support multiple resources at and below the core-level
Problem: The recursive parser throws an error whenever multiple
resources exist in a jobspec. The only exception is at the leaf-level,
where multiple resources are allowed as long as the first one is a core.
This prevents still actionable jobspecs like
`socket->slot->(core->PU,gpu)`. It also prevents
`socket->slot->(gpu,core)`.

Solution: Support multiple resource at or below the core-level.  Below
the core-level is handled by the fact that we do not recurse beyond the
core-level.  Multiple resources at the core-level is handled by
iterating over resources whenever an array of resources is encountered
and explicitly checking for "core" resources in the array.  If none is
found, error out.  If one is found, use it and continue as before in
calculating cores-per-slot.