build: remove recursive Makefile.am in python bindings
Combine the Makefile.am files in src/bindings/python/flux and
src/bindings/python/flux/core into a single Makefile.am. Use nobase_
prefix so that all Python source files can be listed in the same
rule. The special `fluxpycoredir` automake variable is no longer
needed, so it can be dropped.

This should make the bindings Makefiles a bit easier to maintain
by reducing duplication and the number of files to edit when new
Python source files are added.