broker/state_machine: add disconnect handler
Problem: if the monitor module or another user of the
state-machine.monitor or state-machine.quorum-monitor streaming
RPCs disconnect, responses continue to be sent to that client.

Implement a disconnect handler that purges clients that have
disconnected.

Note this does not apply to requests from TBON peers, as they
do not trigger disconnect messages when they shut down (at least for now).