broker: drop -g,--shutdown-grace option
Problem: the timeout over the shutdown state will need to be
reworked given that the purpose and semantics of the SHUTDOWN
state have changed.  It is not currently used in any case and
probably should be set another way.

Drop this option, although to avoid breaking sched builds,
accept it and emit a warning that it's deprecated.

Update docs and test usage.