python: add methods to get JobInfo from list_jobs/ids
Problem: The Python flux.job.job_list() and job_list_id() return
the JSON payload of the job-info list RPC directly, which is not
as user friendly returning JobInfo objects.

Add new "getters" for the JobListRPC and JobListIdRPC classes to
return a list (actually a generator) of JobInfo objects, or a single
JobInfo object respectively.