python: allow future.get_flux() to return None
Problem: flux_future_get_flux(3) returns NULL with EINVAL when the
future has not yet has flux_future_set_flux(3) called on it. This
makes it difficult for a caller to check if a future has a flux
handle installed.

Catch EINVAL in Future.get_flux() and return None if caught.