python: add python WaitAllFuture class
Add a container which can synchronously wait on a set of child futures
to be fulfilled.

WaitAllFuture subclasses Future and provides an additional "push"
method to push child futures into the composite future.

Calling .get() or .wait_for() or .then() on a WaitAllFuture which
wait until *all* child futures have been fulfilled with either success
or error.