python: JobList: support listing by jobids
Problem: The interface for listing jobs from python is different
when listing all jobs with filtering vs listing jobs by jobid.

Introduce a new JobListIdsFuture class which simulates the interface
of JobListRPC, but is implemented via a WaitAllFuture. The JobList
class will return one of these composite futures when a list of ids
is requested rather than a set of all jobs with filtering, since the
jobid listing must be split into multiple RPCs.

The JobListIdsFuture will be fulfilled once RPCs for each jobid are
fulfilled, either by success or error. Errors are accumulated on
an errors list for later access by the caller. This allows the caller
to potentially print information about jobs for which the RPCs were
successful, while still emitting all RPC errors.