rc: load sched-simple after other rc1s
Problem: fluxion rc1 scripts have to unload sched-simple because
flux-core rc1 loads it unconditionally.

Change rc1 so that sched-simple is loaded only if an external
rc1 script did not load a module that registered the 'sched' service.

In addition, if a scheduler is not loaded, check the FLUX_SCHED_MODULE
environment variable:
- if set to "none", don't load anything
- if set to any other string, load a module by that name
- if unset, load sched-simple