python/job: handle empty string conversions
Annotations that do not exist are automatically converted to empty
strings.  This is generally fine, except for output conversions that
assume that the value passed in is always an integer.

For the D, F, and H conversion flags, if an empty string is
passed in for input, output an empty string instead.

For the d conversion flag, output the epoch datetime (i.e. 0.0).

Update documentation.

Fixes #3161