job-manager: add validate RPC
Add job-manager.validate RPC that can look up a job ID and
return the basic info that the job manager keeps for it:
  userid - job owner
  t_submit - submit timestamp
  priority - job priority
  state - job state
  flags - submit flags

Inactive jobs are handled too.  If the job is not found in the
job manager's active job hash, its eventlog is fetched from the
KVS and replayed.

If the KVS lookup fails with ENOENT, the RPC returns with ENOENT
and an "invalid job ID" human-readable message.  This response is
definitive since the submit RPC does not return a job ID to the
user until the job manager is aware of it.

This RPC is currently open to guests.

Fixes #3191