modules/job-info: handle annotations race
There is a small chance that a job-annotation could arrive to the
job-info module before the job-info module is aware of a jobs
existance.

In the event this occurs, save the annotation off to a new
"early_annotations" hash.  When the job-info modules becomes
aware of the job via a job-state transition, load the annotation
out of the hash.

Add basic racy annotations test.

Fixes #3165