broker/log: add log-stderr-mode attribute
Problem: when a broker exits because its upstream has shut down,
there are no flux logs in the systemd journal; instead, the logs
for brokers of rank > 0 appear in the journal on rank 0.  This may
be surprising when a sys admin is trying to debug why a broker is
not running on a particular node.

Add a new log-stderr-mode attribute which may have the following
values:

leader (default)
  Only rank 0 emits log message to stderr, on behalf of all ranks

local
  Each rank emits its own messages to stderr.

Set log-stderr-mode=local in the systemd unit file.

Update flux-broker-attributes(7) man page.

Fixes #3189