resource: fail resource.acquire if hwloc reload fails
Problem: if the resource module detects that flux hwloc reload
has failed, it does not respond with error to a pending
resource.acquire request from the scheduler.

Since the scheduler may be waiting synchronously for the first
resource.acquire RPC, this may result in the scheduler becoming
unresponsive.

Resource discovery already posts the hwloc-discover-finish
event to the resource eventlog if the subprocess returns failure,
and arranges for discover_get() to return NULL in that case.
It is a small change to simply fail the resource.acquire request
when discover_get() returns NULL.