resource: abort resource discovery on shutdown
Problem: the scheduler can potentially hang in a synchronous RPC
to resource.acquire if the instance is shut down before resource
discovery has started or while it is in progress.

Subscribe to the broker shutdown message and
- kill -9 flux hwloc reload if in progress
- post hwloc-discover-finish event if not started

Either path will resut in hwloc-discover-finish being posted.
The acquire logic will call discover_get() in response to the
hwloc-discover-finish event, which will return NULL and cause
resource.acquire to return an error response.

Fixes #3197