modules/resource: load module on all broker ranks
Problem: the resource module is only loaded on rank 0, so
it is entirely dependent on external services for monitoring
execution target availability.

Set up the resource module so that it is loaded on all
broker ranks.  Currently, only the 'monitor' subsystem is
initialized, and then disabled.

This will provide a framework for adding monitoring capability
within the resource module.

In some sharness tests where it will be needed later, add the
'monitor-force-up' resource module option.  This is currently
a no-op.