resource/monitor: use internal hello/goodbye RPCs
Problem: the resource module uses the broker state machine quorum
monitor to determine when resources are online, however there is no
mechanism in the state machine to inform when nodes go offline.

Perform the monitoring within the resource module, which is now loaded
on all ranks.  In initialization, send a "hello" RPC upstream.
The hello idsets are batched and aggregated up the TBON to rank 0.
When the module is unloaded, send a "goodbye" RPC upstream.
The goodbye idsets are batched and aggregated up the TBON to rank 0.

This simple system allows "systemctl stop flux" on a compute node to
remove the node from the pool of available resources automatically.

Add an option 'monitor-force-up' for testing that short-circuits
monitoring, forcing all ranks to appear up in the initial 'reload' event.

Fixes #3187