broker/state_machine: avoid unnecessary quorum noise
Problem: quorum is no longer used by resource module, so there
is no need for nodes that are not required for quourm to check in.

Don't report if rank is not a member of 'want' (broker.quorum) idset.

Mark quorum as full if 'want' == 'have' on rank 0, as opposed to
check that 'want' is a subset of 'have'.

Drop test_idset_subset() helper.