python: allow Future and RPC get to be interrupted
Problem: The Python interpreter installs a signal handler for SIGINT
so that it can raise KeyboardInterrupt when a user hits Ctrl-C.
Unfortunately, when calling into a C function, the exception cannot be
raised until execution returns to Python. For Flux calls that may sleep
in the reactor, this effectively makes those calls uninterruptible,
which is not very user friendly.

Add a decorator function to flux.util, @interruptible, which will reset
the SIGINT handler to SIG_DFL for the duration of the wrapped function.
Apply this decorator to methods of Future and RPC classes that make
blocking calls into the C API, namely .get() and .wait_for(), so that
utilities blocked in these methods can be interrupted by Ctrl-C at
the command line.

Fixes #3186