python: make JobEventWatchFuture get_event() interruptible
Problem: The JobEventWatchFuture get_event() method is not interruptible
by Ctrl-C while it is blocked in the C API at the call to
`RAW.event_watch_get(self.pimpl, result)`, and the @interruptible
decorator can't be used since this is a direct cffi call.

To allow the blocking part of the call to be interrupted, use
future.wait_for() to ensure the future is ready before calling the
event_watch specific get. This function is already wrapped in
@interruptible decorator.

Add a call to future.wait_for() to ensure the future is ready before
calling RAW.event_watch_get() in the JobEventWatchFuture