hwloc: add printing of num GPUs to `flux hwloc info`
Problem: On GPU-enabled clusters, `flux hwloc info` only prints `X
Machines, Y Cores, Z PUs`.  This makes it more difficult to debug
unsatisfiable jobspecs and if the hwloc that Flux is using was
configured with CUDA support.

Solution: add counting of the number of GPUs to `cmd_info` so that the
output on GPU-enabled clusters is now `X Machines, Y Cores, Z PUs, W GPUs`.
The output on non-GPU clusters is unaffected.