broker: fix pmi bootstrap time logging
Problem: elapsed PMI bootstrap time is not logged

Elapsed time in the PMI bootstrap is logged with flux_log() before
the logging system is initialized, so the message is dropped.

Tack the boot time onto the existing "boot:" message that's emitted
directly to stderr if the broker --verbose option was specified.