README.md: advise luarocks for lua-posix error
Problem: having instructions to manually drop a symlink in the
root file system to fix lua-posix as part of buiding flux is not
exactly advising best practice.

Confirmed: lua-posix is _still_ broken on bionic.
It is not an issue on focal.

Instead of the symlink hack, advise installing lua-posix with luarocks
on bionic as was done in the bionic Dockerfile, and reference the
Launchpad issue.

Hopefully people are moving off of bionic now that there is a focal LTS.