t/t2211-job-manager-events: Refactor event checks
Refactor how events are checked for in preparation for future tests.

Instead of simple checking that an event name has occurred in a file,
ensure that event name has occurred for a specific jobid.

In addition, instead of using waitfile to check that a string has
occurred in a file, ensure that the event has occurred for a specific
jobid before continuing.