t/t2211-job-manager-events: Refactor event checks
Problem: Tests in t2211-job-manager-events check for events via a
simple grep for the event name.  This can be a problem in the future
when multiple jobid's events could exist in the event stream.
Similarly, the waitfile script is used to check for a generic pattern,
independent of jobid.

Solution: Refactor tests to check that an event name has occurred for a
specific jobid.  Add a new function to wait for an event name tied to
a specific jobid.