job-manager: Support EVENT_JOURNAL_ONLY flag
Problem: the event_job_post_pack() function presently writes
all events to the job eventlog in the KVS.  But in the future
we would like some events to only be posted to the event journal
and not the KVS job eventlog.

Solution: Support a new EVENT_JOURNAL_ONLY flag, to indiciate that
this specific event should only go out on the event journal.