job-manager: Support EVENT_INFO_ONLY flag
Problem: the event_job_post_pack() function presently writes
all events to the job eventlog in the KVS.  But in the future
we would like some events to only be sent out on the event stream
and not to the job eventlog.

Solution: Support a new EVENT_INFO_ONLY flag, to indiciate that
this specific event should only go out on the event stream.