job-manager: send journal events immediately
Problem: listener responses are generated in event_batch_destroy(),
which may come after a KVS commit.  During the time in which we
are waiting for the KVS commit to complete, its possible
additional journal events could be generated and sent in the
next batch, leading to journal events being sent out of order.

Solution: Send all journal events immediately, not waiting
for the batch to occur.  This ensures that all journal events are
sent in order to listeners.

Fixes #3245