job-manager: send events stream immediately
Problem: listener responses are generated in event_batch_destroy(),
which may come after a KVS commit.  During the time in which we
are waiting for the KVS commit to complete, its possible
additional event stream responses could be appended into the
older batch, when they should belong in the next batch.

Solution: Send all events for the events stream immediately in
event_batch_commit(), do not wait for the KVS commit to complete before
doing so.  This ensures that all events in the stream are sent in order
to listeners.

Note that if we were generate listener responses in event_batch_commit()
and send them later in event_batch_destroy() this could create a scenario
in which events could be sent out of order.  This race occurs due to the
fact the KVS commit can be slow, leading to a later batch (perhaps not having
a KVS commit) completing before an early batch.  Therefore the solution must
be to send all of the responses immediately.

Fixes #3245