job-manager: fix events stream race & bug
Problem: listener responses are generated in event_batch_destroy(),
which may come after a KVS commit.  During the time in which we
are waiting for the KVS commit to complete, its possible
additional event stream responses could be appended into the
older batch, when they should belong in the next batch.

Solution: Generate listener responses in event_batch_commit()
and not event_batch_destroy().