job-manager: fix zlist_freefn call
Problem: zlist_freefn() sets the destructor on a list item, searching for the
item starting from the beginning of the list.  As an optimization, the at_tail
parameter can be set if the item was recently appended to the list, so the
entirely list doesn't have to be scanned.

In the events_handle_request() function, we append to the listeners list, then
call zlist_freefn() with at_tail=false, which is inefficient.

Solution: Set at_tail=true here.