job-manager: Add sequence nums to journal events
Problem: Some users of the events journal stream may also read
events via the job eventlog in the KVS.  It may be difficult to
determine if an event from the journal has already been read from
the eventlog.

Solution: Add sequence numbers to journal events to indicate the
event sequence number for a job.  Users will be able to use the
sequence number to determine if an event from the eventlog has
already been read / processed.