job-info: check if annotations are json null
Problem: annotations posted in an alloc event may be the json null value,
if all annotations have been cleared.  However, this is not checked for
when the job-info module is restarted and an eventlog is being read for
re-initialization of job list state.

Solution: Check if annotations are equivalent to json null before setting
the job->annotations field.