job-info: Refactor eventlog_inactive_finish()
Problem-1: The eventlog_inactive_finish() could be confusing to readers
b/c it may imply it is related to the "finish" event in the eventlog.

Problem-2: eventlog_inactive_finish() returns 0 all of the time, it cannot
produce an error.

Solution: Rename the function to eventlog_inactive_complete().
In addition, return void, as the function cannot produce an error.  Adjust
callers accordingly.