python: fix signal handler management in threads
Shield the signal handler code from being invoked in subthreads (i.e., not the
main thread), Python does not like that.  Since the KeyboardInterrupt which
codes wants to enable arrives in the MainThread anyway, this should be fine.
Subthreads are difficult to terminate then though, it is adviseable to declare
them as deamon threads so that MainThread interruptions do not hang on pending
thread activities.

This fixes #3264.