resource: add container and transition to Rv1
Add inventory.c, which is a container for resources known to the
resource module.  It adds methods for populating resources from
configuration (file pointed to from TOML), from R in the enclosing
instance, and from discovery.

The resource module commits R to the KVS as 'resource.R'.  This is used
on a restart unless the resources are defined via TOML config, and then
it is ignored.

It expects R version 1 instead of the by_rank resource object format.
Other parts of the resource module such as acquire.c and discover.c are
adjusted accordingly.

Fake resources may be loaded in test using the resource.reload RPC.

Add an resource.get-xml RPC (rank 0 only) which fetches a fixed
length array from 0...size-1 of XML strings.  If XML is not yet
in inventory, block until it arrives.

Sched-simple is adjusted to acquire resources in Rv1 form.

Tests are updated.