sched-simple: use librlist to set Rv1 timelimits
Problem: sched-simple add starttime and expiration keys to the Rv1 object
after it is encoded from the "struct rlist" allocated to the job.
This is unnecessary now that librlist supports these keys.

Set starttime and expiration if necessary in the 'struct rlist' before
encoding it to Rv1, and remove previous code that explicitly encoded
these values after the fact.