flux-mini: print jobid with -v option to flux-mini alloc
Problem: There is no option to emit the submitted jobid with flux-mini
alloc. This may be useful for users who want to query a pending jobid
in another window, or for cases where the command is somehow scripted.

Add a `-v, --verbose` option which, similar to flux-mini run, will print
the submitted jobid on stderr before moving on with program operation.

Fixes #3274