rc1: increase local connector retries
Problem: valgrind test occasionally fails with 30 retries
on the local connector.

The timeout is in rc1 on the first access to the broker.
Raise it 40 retries.

N.B. the retry time starts at 0.016s and doubles each time until
it reaches 2s, then it's 2s each up to the maxmimum.  So times are:
  0.016, 0.032, 0.064, 0.128, 0.256, 0.512, 1.024, 2, 2, 2, ....
Therefore 30 retries is about 26s and 40 retries is about 46s.