broker: clear all log entries with log.clear
Refactor the log.clear request to always clear all entries.

As a consequence, we remove the `seq` argument in the RPC and adjust
callers appropriately.  The logbuf_clear() function is also removed
as a result.

As the RPC no longer expects arguments, remove an empty payload request
test, which is no longer a failure.