broker: update log.dmesg to use rpc streaming
Update the log.dmesg RPC to use modern flux RPC streaming, rather
than the old implementation that required callers to request
an increasing series of sequence numbers.  This not only modernizes
the RPC but should improve performance as repeated iteration of the
log buf list will no longer be necessary.

As a result of this change, the `seq` argument to log.dmesg is no longer
necessary.  It has been removed and all callers and associated code have
been adjusted.  In addition, all of the "sleepers" code in broker/log.c
has been removed as it is no longer necessary.

Fixes #3212