libsubprocess: do not allow ref/unref in hooks
Problem: In the pre_exec and post_fork hooks, the user has the potential
to destroy the subprocess, which can lead to potential segfaults later on.  We
should treat hooks as only having "read only" access to flux_subprocess_t
contents.

Solution: Do not allow the user to call flux_subprocess_ref() or
flux_subprocess_unref() while in hooks.

Fixes #2225