broker: set security.owner attribute
Problem: there is no straightforward way for flux API users to know
whether they are the instance owner of a given Flux instance or not.

Add the 'security.owner' broker attribute, set to the userid that is
runing the Flux broker.  This is an immutable attribute so it cannot
be changed by users, and it is automatically cached on first use
by `flux_attr_get()`.