libjob: sign with 'none' if instance owner
Problem: If flux is built --with-flux-security, all jobs are signed
with the default signing mechanism, even if the job is submitted as
the instance owner.

If munge is the default mech, this means the munge daemon has to be running
for the instance owner to submit jobs, which creates practical complications
for testing, e.g. running 'make check' in the rpmbuild environment.
In addition, the overhead of communicating with the munge daemon, although
minor, is unnecessary in that case.

Modify flux_job_submit() to fetch the security.owner broker attribute.
If the submitting user == instance owner, force the 'none' mech.
The job-ingest module will accept the job because the authenticated user
matches the instance owner.  The execution system will execute the job
because the IMP is bypassed for the instance owner.

Fixes #3305