testsuite: fix unlikely core idsets in resource tests
Problem: The t2310-resource-module.t and t2315-resource-system.t
tests set an "unlikely" core idset for testing of 0-36 and 0-48
in order to force ranks to be drained. However, ranks will not
be drained on systems with > 37 cores since nodes exceed the
configured cores.

Choose an extremely unlikely core idset, 0-1048, to make these tests
less brittle.

Fixes #3312