shell: gpubind: fix CUDA_VISIBLE_DEVICES when no GPUs are allocated
Problem: When a job has no GPUs allocated, the shell gpubind plugin
sets CUDA_VISIBLE_DEVICES=-1 to ensure that all GPUs on a system
are hidden from the job (in an advisory manner). However, the call
to flux_shell_setenvf(3) does not set the overwrite flag, so if
an existing value for CUDA_VISIBLE_DEVICES is set in the current
environment, this value is used instead, leading jobs to discover
GPU devices which have not be granted.

Add the overwrite flag to flux_shell_setenvf(3) to fix the issue.

Fixes #3316