resource: drain: support drain by hostname
Problem: The resource drain protocol supports draining ranks by
rank id only. This may be inconvenient and error prone for system
administrators which may be used to operating on hostnames.

Support "targets" instead of a strict idset in the resource module
drain protocol. A set of targets may be encoded as either an RFC22
idset string or RFC29 hostlist. If a hostlist is specified, all
ranks matching the hostname or hostnames specified will be drained.
(There may be a many-to-one relationship between ranks and hostnames
on a system with > 1 broker per node)