resource: exclude: support exclusion by hostname
Problem: The resource module target exclusion only works with ranks,
but it may be more convenient for sysadmins or users to exclude
targets by hostname.

Support exclusion by "targets" instead of strict idset as was done
in the drain protocol. A targets list can be either a rank idset or
RFC29 hostlist.