resource: use 'targets' instead of 'idset' in drain protocol
Problem: The name of target list in the drain protocol and flux-resource
drain command is still "idset", but this may be confusing since the drain
protocol takes a target list which is either an idset or list of hosts.

Update the drain protocol and flux-resource drain,undrain commands to
specify "targets" instead of "idset".