build: python: fix unnecessary generation of cffi sources
Problem: The cffi emit_c_code() function does not update the target
if it believes the generated .c file is already up-to-date. This causes
make confusion because the file is never updated if one of its other
prerequisites changes (e.g. Makefile), so the rule will continually
be rerun on each make invocation.

Ensure the target file mtime is updated in the *_build.py scripts by
"touching" the file after ffi.emit_c_code() is complete. This way make
won't unnecessarily re-run the rule to update the target.