resource/drain: retain drain details in memory
Problem: resource status query will need to return drain reason
and timestamp, but we don't keep those in memory.

Add some additional data structures to track drain reason and
timestamp.

This changes the format of the 'resource-init' event in the eventlog
so that it includes the timestamp and reason for all drained nodes.
The design of the resource eventlog replay is that all events prior
to the _last_ restart are pruned, so 'resource-init' needs to contain
a summary that the next startup and begin with.  So changing the event
to include that info is the only way to preserve it across a restart.

In addition, due to the internal changes in drain.c, the drain_get()
function changed so that the caller must free the result.

Update users.