resource/drain: don't bother with undrain recovery
Problem: if undrain_cb() fails to post its event, we try to recover
by removing the idset from the drain set.  Now that there will be
more state to undo, it seems like an unworthy effort, and also
different from the error handling in drain_cb().

Drop this recovery-on-error code.