job-exec: fix truncated input to IMP
Problem: The job-exec module artificially constrains the stdin buffer
to 8K for remote processes when invoking the IMP (i.e. multi-user jobs).
The result is that the input to the IMP, containing the signed jobspec
J, is truncated at 8K and thus larger jobspecs cannot be run in multiuser
mode.

Remove the offending line that reduces the max buffer size so that the
default max is used ~4M. Signed jobspec above this size will still be
truncated, but this will need to be fixed in the future.

Fixes #3347