job-exec: log fatal error on failure to write all IMP input
Problem: In multiuser mode the job-exec module writes input to all
IMP processes using bulk_exec_write(), but the return code of this
function is ignored, making it difficult to diagnose problems due to
a failure writing input to the IMP.

Check the return code from bulk_exec_write() and log a useful fatal
error message for the job.

Also cast the return of bulk_exec_close() to void to indicate we're
ignoring this error on purpose.