resource/inventory: use rutil_idkey to reduce XML
Problem: XML is "reduced" using a rank-indexed object, but we
don't bother trying to consolidate ranks with identical XML.

Reducing identical XML should be possible when there are multiple
brokers per node.  Use the new rutil_idkey_*() functions to reduce
and manipulate idset-indexed XML objects.

Instead of storing the collected XML as a rank-indexed array,
store the reduced form and convert to the array on demand when
responding to the resource.get-xml RPC.