build: t/Makefile.am: reorder libflux-core.la in test builds
Problem: On some systems, when compiling --with-flux-security, some
test programs in t/* will link against a system-installed libflux-core
instead of the builddir version. This seems to be a libtool bug.

A workaround appears to be ensuring that libflux-core.la appears
after all static/conveniece libraries in test_ldadd, though it is
not clear why that works.